latest

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -10,13 +10,13 @@
 int main(void)
 {
 	int count = 2;
-	int64_t i = 1;
-	int64_t j = 2;
+	long long i = 1;
+	long long j = 2;
 
 	printf("1, 2, ");
 	while (count != 98)
 	{
-		int64_t new = i + j;
+		long long new = i + j;
 
 		i = j;
 		j = new;
